Define basic build pipeline
This defines a minimal application script along with a build pipeline
utilising `bash-tpl` to include dynamic information in the final
distributed application.

The build number is dynamically determined from the next version plus
some prerelease and build information to indicate it is still in
progress.

The README has also been updated to describe this build process.